Remove retired ItemsRepeater control
