+++ Source/LibationAvalonia/LibationAvalonia.csproj
@@ -77,7 +77,6 @@
 		<PackageReference Include="Avalonia.Diagnostics" Version="11.3.0" Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'" />
 		<PackageReference Include="Avalonia" Version="11.3.0" />
 		<PackageReference Include="Avalonia.Controls.DataGrid" Version="11.3.0" />
-		<PackageReference Include="Avalonia.Controls.ItemsRepeater" Version="11.1.5" />
 		<PackageReference Include="Avalonia.Desktop" Version="11.3.0" />
 		<PackageReference Include="Avalonia.ReactiveUI" Version="11.3.0" />
 		<PackageReference Include="Avalonia.Themes.Fluent" Version="11.3.0" />
